requirements.txt
pymongo
faker
python-dotenv
neo4j
matplotlib
seaborn
pandas
flask
flask-cors